// src/enums.ts
export enum dayOfWeek {
  Monday = 'Monday',
  Tuesday = 'Tuesday',
  Wednesday = 'Wednesday',
  Thursday = 'Thursday',
  Friday = 'Friday',
  Saturday = 'Saturday',
  Sunday = 'Sunday'
}
export enum typeOfDate {
  Daily = 'Daily',
  Weekly = 'Weekly',
  Monthly = 'Monthly',
  Once = 'Once',
  Custom = 'Custom'
}